IP-20 updated the model from result set method and readLatest method in the order DAO by using a OrderItem DAO

+++ src/test/resources/sql-schema.sql
@@ -2,14 +2,14 @@ DROP TABLE IF EXISTS `customers`;
 DROP TABLE IF EXISTS `items`;
 
 CREATE TABLE IF NOT EXISTS `customers` (
-    `id` INT(11) NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `first_name` VARCHAR(64) NOT NULL,
     `surname` VARCHAR(64) NOT NULL,
     PRIMARY KEY (`id`)
 );
 
 CREATE TABLE IF NOT EXISTS `items` (
-    `id` INT(11) NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `type` VARCHAR(40) NOT NULL,
     `name` VARCHAR(40) NOT NULL,
     `description` VARCHAR(256),
